+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Released on 2023-07-27: :package: `p2panda-js` and :package: `p2panda-rs`
 ### Fixed
 
 - Handle parsing empty pinned relation lists from `PlainOperation` [#489](https://github.com/p2panda/p2panda/pull/489) `rs`
+- Fix missing bytes field validation in schema field definition [#519](https://github.com/p2panda/p2panda/pull/519) `rs`
 
 ### Lion on the street